+++ quiz_quest/lib/app/features/quiz_pages/tv_quiz_pages/easy_tv_quiz_page/easy_question_quiz_page.dart
@@ -37,7 +37,9 @@ import 'package:quiz_quest/app/features/quiz_pages/sports_quiz_pages/cubit/sport
 import 'package:quiz_quest/app/features/quiz_pages/sports_quiz_pages/easy_sport_quiz_page/easy_sport_lost_life_quiz_page.dart';
 import 'package:quiz_quest/app/features/quiz_pages/sports_quiz_pages/easy_sport_quiz_page/resume_easy_sport_question_quiz_page.dart';
 import 'package:quiz_quest/app/features/quiz_pages/tv_quiz_pages/cubit/tv_cubit.dart';
+import 'package:quiz_quest/app/features/quiz_pages/tv_quiz_pages/easy_tv_quiz_page/easy_tv_answer_button.dart';
 import 'package:quiz_quest/app/features/quiz_pages/tv_quiz_pages/easy_tv_quiz_page/easy_tv_lost_life_quiz_page.dart';
+import 'package:quiz_quest/app/features/quiz_pages/tv_quiz_pages/easy_tv_quiz_page/easy_tv_question_widget.dart';
 import 'package:quiz_quest/app/features/quiz_pages/tv_quiz_pages/easy_tv_quiz_page/resume_easy_tv_question_quiz_page.dart';
 import 'package:quiz_quest/app/injection_container.dart';
 
@@ -47,8 +49,7 @@ class EasyQuestionTvQuizPage extends StatefulWidget {
   });
 
   @override
-  State<EasyQuestionTvQuizPage> createState() =>
-      _EasyQuestionTvQuizPageState();
+  State<EasyQuestionTvQuizPage> createState() => _EasyQuestionTvQuizPageState();
 }
 
 final controller = CountDownController();
@@ -137,8 +138,7 @@ class _EasyQuestionTvQuizPageState extends State<EasyQuestionTvQuizPage> {
 
     return Scaffold(
       body: BlocProvider(
-        create: (context) => getIt<TVCubit>()
-          ..getEasyTVCategory(),
+        create: (context) => getIt<TVCubit>()..getEasyTVCategory(),
         child: BlocListener<TVCubit, TVState>(
           listener: (context, state) async {
             if (state.status == Status.error) {
@@ -285,13 +285,12 @@ class _EasyQuestionTvQuizPageState extends State<EasyQuestionTvQuizPage> {
                           shrinkWrap: true,
                           itemBuilder: (context, _) {
                             final questionNumber = currentIndex + 1;
-                            final questionNumbers =
-                                tvQuizModel.results.length;
+                            final questionNumbers = tvQuizModel.results.length;
 
                             return Column(
                               mainAxisAlignment: MainAxisAlignment.spaceAround,
                               children: [
-                                QuestionWidget(
+                                EasyTVQuestionWidget(
                                   question: tvQuizModel
                                       .results[currentIndex].question,
                                 ),
@@ -337,7 +336,7 @@ class _EasyQuestionTvQuizPageState extends State<EasyQuestionTvQuizPage> {
                                 for (int index = 0;
                                     index < currentAnswers.length;
                                     index++) ...[
-                                  AnswerButton(
+                                  EasyTVAnswerButton(
                                     isTimeUp: isTimeUp,
                                     duration: duration,
                                     isButtonDisabled: (value) {
@@ -349,8 +348,7 @@ class _EasyQuestionTvQuizPageState extends State<EasyQuestionTvQuizPage> {
                                       setState(() {
                                         isButtonClicked = value;
                                         if (value = currentAnswers[index] ==
-                                            tvQuizModel
-                                                .results[currentIndex]
+                                            tvQuizModel.results[currentIndex]
                                                 .correctAnswer) {
                                           ringColor = Colors.green;
                                         } else {
@@ -365,8 +363,7 @@ class _EasyQuestionTvQuizPageState extends State<EasyQuestionTvQuizPage> {
                                       });
                                     },
                                     isCorrectAnswer: currentAnswers[index] ==
-                                            tvQuizModel
-                                                .results[currentIndex]
+                                            tvQuizModel.results[currentIndex]
                                                 .correctAnswer
                                         ? true
                                         : false,
@@ -398,8 +395,7 @@ class _EasyQuestionTvQuizPageState extends State<EasyQuestionTvQuizPage> {
                                                 ? () {
                                                     setState(() {
                                                       if (currentIndex ==
-                                                          tvQuizModel
-                                                                  .results
+                                                          tvQuizModel.results
                                                                   .length -
                                                               1) {
                                                         Navigator.of(context)
@@ -440,8 +436,7 @@ class _EasyQuestionTvQuizPageState extends State<EasyQuestionTvQuizPage> {
                                               backgroundColor: Colors.black,
                                             ),
                                             child: Text(currentIndex ==
-                                                    tvQuizModel
-                                                            .results.length -
+                                                    tvQuizModel.results.length -
                                                         1
                                                 ? 'Show your results'
                                                 : 'Next Question ➔'),
@@ -458,151 +453,6 @@ class _EasyQuestionTvQuizPageState extends State<EasyQuestionTvQuizPage> {
                 ),
               );
             },
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class QuestionWidget extends StatelessWidget {
-  const QuestionWidget({
-    required this.question,
-    super.key,
-  });
-
-  final String question;
-
-  @override
-  Widget build(BuildContext context) {
-    final modifiedQuestion = question
-        .replaceAll('&quot;', '')
-        .replaceAll('&#039;', '')
-        .replaceAll('&aacute;', '')
-        .replaceAll('&ntilde;', '')
-        .replaceAll('&amp;', '')
-        .replaceAll('&rsquo;', '');
-
-    return Center(
-      child: Text(
-        modifiedQuestion,
-        style: GoogleFonts.aBeeZee(
-            fontSize: 24, color: Colors.white, fontWeight: FontWeight.bold),
-        textAlign: TextAlign.center,
-      ),
-    );
-  }
-}
-
-class AnswerButton extends StatefulWidget {
-  AnswerButton({
-    required this.answer,
-    required this.controller,
-    required this.isCorrectAnswer,
-    required this.colorFunction,
-    required this.isButtonClicked,
-    required this.isButtonDisabled,
-    required this.textcolor,
-    required this.index,
-    required this.duration,
-    required this.isTimeUp,
-  });
-
-  final String answer;
-  final CountDownController controller;
-  final bool isCorrectAnswer;
-
-  int duration;
-  bool isTimeUp;
-  Function(Color, int) colorFunction;
-  final Function(bool) isButtonClicked;
-  final Function(bool) isButtonDisabled;
-  Color textcolor;
-  final int index;
-
-  @override
-  State<AnswerButton> createState() => _AnswerButtonState();
-}
-
-class _AnswerButtonState extends State<AnswerButton> {
-  void onPressed() {
-    if (isButtonDisabled) {
-      return;
-    }
-
-    widget.controller.pause();
-
-    if (widget.isCorrectAnswer) {
-      widget.colorFunction(Colors.green, widget.index);
-      easyTvGoodAnswers += 1;
-    } else {
-      widget.colorFunction(Colors.red, widget.index);
-      easyTvBadAnswers += 1;
-      if (easyTvBadAnswers == 3) {
-        context.read<TVCubit>().updateEasyTVPoints(easyTvGoodAnswers);
-        Navigator.of(context).push(
-          MaterialPageRoute(
-            builder: (context) =>
-                EasyTvLostLifePage(goodAnswers: easyTvGoodAnswers),
-          ),
-        );
-      }
-    }
-    setState(() {
-      widget.textcolor = widget.isCorrectAnswer ? Colors.green : Colors.red;
-    });
-    widget.isButtonClicked(true);
-    widget.isButtonDisabled(true);
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final modifiedAnswer = widget.answer
-        .replaceAll('&quot;', '')
-        .replaceAll('&#039;', '')
-        .replaceAll('&aacute;', '')
-        .replaceAll('&ntilde;', '')
-        .replaceAll('&amp;', '')
-        .replaceAll('&rsquo;', '')
-        .replaceAll('&ocirc;', '');
-
-    return Container(
-      margin: const EdgeInsets.symmetric(horizontal: 20),
-      decoration: BoxDecoration(
-        gradient: const LinearGradient(
-          colors: [
-            Color.fromRGBO(11, 22, 65, 1),
-            Color.fromRGBO(9, 77, 203, 1),
-          ],
-          begin: Alignment.centerLeft,
-          end: Alignment.centerRight,
-        ),
-        borderRadius: const BorderRadius.all(
-          Radius.circular(6.0),
-        ),
-        boxShadow: [
-          BoxShadow(
-            color: Colors.black.withOpacity(0.1),
-            spreadRadius: 4,
-            blurRadius: 10,
-            offset: const Offset(0, 3),
-          )
-        ],
-      ),
-      child: ElevatedButton(
-        onPressed: onPressed,
-        style: ElevatedButton.styleFrom(
-          minimumSize: const Size.fromHeight(50),
-          backgroundColor: Colors.transparent,
-          shadowColor: Colors.transparent,
-        ),
-        child: Text(
-          modifiedAnswer,
-          style: GoogleFonts.aBeeZee(
-            fontSize: 24,
-            color: widget.isTimeUp && widget.isCorrectAnswer
-                ? Colors.green
-                : widget.textcolor,
           ),
         ),
       ),
