+++ quiz_quest/lib/app/features/quiz_pages/tv_quiz_pages/medium_tv_quiz_page/medium_question_quiz_page.dart
@@ -39,7 +39,9 @@ import 'package:quiz_quest/app/features/quiz_pages/sports_quiz_pages/easy_sport_
 import 'package:quiz_quest/app/features/quiz_pages/tv_quiz_pages/cubit/tv_cubit.dart';
 import 'package:quiz_quest/app/features/quiz_pages/tv_quiz_pages/easy_tv_quiz_page/easy_tv_lost_life_quiz_page.dart';
 import 'package:quiz_quest/app/features/quiz_pages/tv_quiz_pages/easy_tv_quiz_page/resume_easy_tv_question_quiz_page.dart';
+import 'package:quiz_quest/app/features/quiz_pages/tv_quiz_pages/medium_tv_quiz_page/medium_tv_answer_button.dart';
 import 'package:quiz_quest/app/features/quiz_pages/tv_quiz_pages/medium_tv_quiz_page/medium_tv_lost_life_quiz_page.dart';
+import 'package:quiz_quest/app/features/quiz_pages/tv_quiz_pages/medium_tv_quiz_page/medium_tv_question_widget.dart';
 import 'package:quiz_quest/app/features/quiz_pages/tv_quiz_pages/medium_tv_quiz_page/resume_medium_tv_question_quiz_page.dart';
 import 'package:quiz_quest/app/injection_container.dart';
 
@@ -291,7 +293,7 @@ class _MediumQuestionTvQuizPageState extends State<MediumQuestionTvQuizPage> {
                             return Column(
                               mainAxisAlignment: MainAxisAlignment.spaceAround,
                               children: [
-                                QuestionWidget(
+                                MediumTVQuestionWidget(
                                   question: tvQuizModel
                                       .results[currentIndex].question,
                                 ),
@@ -337,7 +339,7 @@ class _MediumQuestionTvQuizPageState extends State<MediumQuestionTvQuizPage> {
                                 for (int index = 0;
                                     index < currentAnswers.length;
                                     index++) ...[
-                                  AnswerButton(
+                                  MediumTVAnswerButton(
                                     isTimeUp: isTimeUp,
                                     duration: duration,
                                     isButtonDisabled: (value) {
@@ -461,147 +463,6 @@ class _MediumQuestionTvQuizPageState extends State<MediumQuestionTvQuizPage> {
   }
 }
 
-class QuestionWidget extends StatelessWidget {
-  const QuestionWidget({
-    required this.question,
-    super.key,
-  });
 
-  final String question;
 
-  @override
-  Widget build(BuildContext context) {
-    final modifiedQuestion = question
-        .replaceAll('&quot;', '')
-        .replaceAll('&#039;', '')
-        .replaceAll('&aacute;', '')
-        .replaceAll('&ntilde;', '')
-        .replaceAll('&amp;', '')
-        .replaceAll('&rsquo;', '');
 
-    return Center(
-      child: Text(
-        modifiedQuestion,
-        style: GoogleFonts.aBeeZee(
-            fontSize: 24, color: Colors.white, fontWeight: FontWeight.bold),
-        textAlign: TextAlign.center,
-      ),
-    );
-  }
-}
-
-class AnswerButton extends StatefulWidget {
-  AnswerButton({
-    required this.answer,
-    required this.controller,
-    required this.isCorrectAnswer,
-    required this.colorFunction,
-    required this.isButtonClicked,
-    required this.isButtonDisabled,
-    required this.textcolor,
-    required this.index,
-    required this.duration,
-    required this.isTimeUp,
-  });
-
-  final String answer;
-  final CountDownController controller;
-  final bool isCorrectAnswer;
-
-  int duration;
-  bool isTimeUp;
-  Function(Color, int) colorFunction;
-  final Function(bool) isButtonClicked;
-  final Function(bool) isButtonDisabled;
-  Color textcolor;
-  final int index;
-
-  @override
-  State<AnswerButton> createState() => _AnswerButtonState();
-}
-
-class _AnswerButtonState extends State<AnswerButton> {
-  void onPressed() {
-    if (isButtonDisabled) {
-      return;
-    }
-
-    widget.controller.pause();
-
-    if (widget.isCorrectAnswer) {
-      widget.colorFunction(Colors.green, widget.index);
-      mediumTvGoodAnswers += 1;
-    } else {
-      widget.colorFunction(Colors.red, widget.index);
-      mediumTvBadAnswers += 1;
-      if (mediumTvBadAnswers == 3) {
-        context.read<TVCubit>().updateMediumTVPoints(mediumTvGoodAnswers);
-        Navigator.of(context).push(
-          MaterialPageRoute(
-            builder: (context) =>
-                MediumTvLostLifePage(goodAnswers: mediumTvGoodAnswers),
-          ),
-        );
-      }
-    }
-    setState(() {
-      widget.textcolor = widget.isCorrectAnswer ? Colors.green : Colors.red;
-    });
-    widget.isButtonClicked(true);
-    widget.isButtonDisabled(true);
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final modifiedAnswer = widget.answer
-        .replaceAll('&quot;', '')
-        .replaceAll('&#039;', '')
-        .replaceAll('&aacute;', '')
-        .replaceAll('&ntilde;', '')
-        .replaceAll('&amp;', '')
-        .replaceAll('&rsquo;', '')
-        .replaceAll('&ocirc;', '');
-
-    return Container(
-      margin: const EdgeInsets.symmetric(horizontal: 20),
-      decoration: BoxDecoration(
-        gradient: const LinearGradient(
-          colors: [
-            Color.fromRGBO(11, 22, 65, 1),
-            Color.fromRGBO(9, 77, 203, 1),
-          ],
-          begin: Alignment.centerLeft,
-          end: Alignment.centerRight,
-        ),
-        borderRadius: const BorderRadius.all(
-          Radius.circular(6.0),
-        ),
-        boxShadow: [
-          BoxShadow(
-            color: Colors.black.withOpacity(0.1),
-            spreadRadius: 4,
-            blurRadius: 10,
-            offset: const Offset(0, 3),
-          )
-        ],
-      ),
-      child: ElevatedButton(
-        onPressed: onPressed,
-        style: ElevatedButton.styleFrom(
-          minimumSize: const Size.fromHeight(50),
-          backgroundColor: Colors.transparent,
-          shadowColor: Colors.transparent,
-        ),
-        child: Text(
-          modifiedAnswer,
-          style: GoogleFonts.aBeeZee(
-            fontSize: 24,
-            color: widget.isTimeUp && widget.isCorrectAnswer
-                ? Colors.green
-                : widget.textcolor,
-          ),
-        ),
-      ),
-    );
-  }
-}
